+++ pkg/sql/codegen/xgboost/codegen_test.go
@@ -24,7 +24,12 @@ import (
 
 func TestTrain(t *testing.T) {
 	a := assert.New(t)
+	tir := mockTrainIR()
+	_, err := Train(tir)
+	a.NoError(err)
+}
 
+func mockTrainIR() *codegen.TrainIR {
 	cfg := &mysql.Config{
 		User:                 "root",
 		Passwd:               "root",
@@ -43,7 +48,7 @@ func TestTrain(t *testing.T) {
 	COLUMN sepal_length, sepal_width, petal_length, petal_width
 	LABEL class
 	INTO sqlflow_models.my_xgboost_model;`
-	ir := &codegen.TrainIR{
+	return &codegen.TrainIR{
 		DataSource:       fmt.Sprintf("mysql://%s", cfg.FormatDSN()),
 		Select:           "select * from iris.train;",
 		ValidationSelect: "select * from iris.test;",
@@ -60,6 +65,4 @@ func TestTrain(t *testing.T) {
 				&codegen.NumericColumn{&codegen.FieldMeta{"petal_length", codegen.Float, "", []int{1}, false, nil}},
 				&codegen.NumericColumn{&codegen.FieldMeta{"petal_width", codegen.Float, "", []int{1}, false, nil}}}},
 		Label: &codegen.NumericColumn{&codegen.FieldMeta{"class", codegen.Int, "", []int{1}, false, nil}}}
-	_, err := Train(ir)
-	a.NoError(err)
 }
